requirements.txt
pytz
requests
ipython

# Django
django==4.0.2
django-timezone-field 
django-cors-headers
django-debug-toolbar
django-environ
django-sesame
whitenoise
django-webpack-loader

# Django 3.0 async requirements
gunicorn
uvicorn
uvloop
httptools

# Databases
psycopg2-binary
dj-database-url

# Distributed tasks
celery
redis
django-celery-beat

# Telegram
python-telegram-bot
telethon
django-extensions
werkzeug

# monitoring
# sentry-sdk
